Clone/requirements.txt
PyQt5
PyPDF2
docx2txt
Pillow
matplotlib
numpy
wordcloud
requests
openai
python-dotenv
eel
